Moving job scheduler into a worker
